frontend/requirements.txt
streamlit
requests
python-dotenv
sounddevice
scipy
numpy
